# docs/openapi.yaml
components:
  schemas:
    health.Response:
      properties:
        stage:
          example: prod
          type: string
        status:
          example: ok
          type: string
        version:
          example: 1.2.3+abc1234
          type: string
      required:
      - stage
      - status
      - version
      type: object
    httpx.Error:
      properties:
        code:
          example: not_found
          type: string
        message:
          example: item not found
          type: string
      required:
      - code
      - message
      type: object
    item.CreateInput:
      properties:
        metadata:
          additionalProperties:
            type: string
          description: Metadata, 20 entries max. Keys 100 chars, values 1000.
          example:
            owner: platform
          type: object
        name:
          description: Name is required, trimmed, 200 chars max.
          example: first item
          maxLength: 200
          type: string
        tags:
          description: Tags, 20 max, 50 chars each.
          example:
          - demo
          - starter
          items:
            maxLength: 50
            type: string
          maxItems: 20
          type: array
          uniqueItems: false
      required:
      - name
      type: object
    item.Item:
      properties:
        createdAt:
          example: "2026-08-27T10:00:00Z"
          format: date-time
          type: string
        id:
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          type: string
        metadata:
          additionalProperties:
            type: string
          example:
            owner: platform
          type: object
        name:
          example: first item
          maxLength: 200
          type: string
        tags:
          example:
          - demo
          - starter
          items:
            maxLength: 50
            type: string
          maxItems: 20
          type: array
          uniqueItems: false
      required:
      - createdAt
      - id
      - name
      type: object
    items.ListResponse:
      properties:
        count:
          type: integer
        items:
          items:
            $ref: '#/components/schemas/item.Item'
          type: array
          uniqueItems: false
      required:
      - count
      - items
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: Shared secret. Deployed stages read it from SSM Parameter Store.
      in: header
      name: X-Api-Key
      type: apiKey
externalDocs:
  description: Source repository and README
  url: https://github.com/AntoinePoisson/go-aws-serverless-starter
info:
  description: |-
    Item resource served by two Lambda functions behind a single HTTP API.

    The /items routes are authenticated with the X-Api-Key header and are
    served by the `api` function. /health and /public/... are open and are
    served by the `public` function. Both sit behind the same API Gateway,
    which is why the public read is namespaced under /public.

    Errors carry a JSON body of the shape {"code": "...", "message": "..."}.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Go AWS Serverless Starter API
  version: "1.0"
openapi: 3.1.0
paths:
  /health:
    get:
      description: |-
        Answers 200 whenever the function runs. The version identifies the
        deployed build; it reads "dev" outside a deployed stage.
      operationId: getHealth
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/health.Response'
          description: OK
      summary: Liveness, stage and version
      tags:
      - health
  /items:
    get:
      description: |-
        A limit outside [1, 100] falls back to 25. The count is the size of
        the page, not a total: there is no cursor and no total count.
      operationId: listItems
      parameters:
      - description: Page size, 1 to 100
        in: query
        name: limit
        schema:
          default: 25
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/items.ListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: invalid_limit
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: internal_error
      security:
      - ApiKeyAuth: []
      summary: List items
      tags:
      - items
    post:
      description: Stores an item and returns it with its generated identifier.
      operationId: createItem
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/item.CreateInput'
              description: Item to create
              summary: body
        description: Item to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/item.Item'
          description: Created
          headers:
            Location:
              description: Path of the created item
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: invalid_body or invalid_input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: unauthorized
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: body_too_large, over 1 MiB
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: internal_error
      security:
      - ApiKeyAuth: []
      summary: Create an item
      tags:
      - items
  /items/{id}:
    delete:
      description: Deleting an identifier that does not exist is a 404, not a no-op.
      operationId: deleteItem
      parameters:
      - description: Item identifier
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted, no body
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: not_found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: internal_error
      security:
      - ApiKeyAuth: []
      summary: Delete an item
      tags:
      - items
    get:
      description: Any non-empty identifier is accepted; an unknown one is a 404.
      operationId: getItem
      parameters:
      - description: Item identifier
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/item.Item'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: not_found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: internal_error
      security:
      - ApiKeyAuth: []
      summary: Read an item
      tags:
      - items
  /public/items/{id}:
    get:
      description: Same payload as GET /items/{id}, without the API key.
      operationId: getPublicItem
      parameters:
      - description: Item identifier
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/item.Item'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: not_found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/httpx.Error'
          description: internal_error
      summary: Read an item, unauthenticated
      tags:
      - public
servers:
- description: Local api function
  url: http://localhost:8080
- description: Local public function
  url: http://localhost:8081
- description: Deployed stage, both functions behind one API Gateway
  url: https://{apiId}.execute-api.{region}.amazonaws.com
  variables:
    apiId:
      default: xxxxxxxxxx
    region:
      default: eu-west-1
tags:
- description: Authenticated CRUD over the item resource.
  name: items
- description: Unauthenticated reads, namespaced under /public.
  name: public
- description: Liveness, stage and version.
  name: health